Server/requirements.txt
fastapi
uvicorn
mysql-connector-python
docker
docker-compose
PyJWT